requirements.txt
# Core dependencies for Phase 1
streamlit==1.28.0
PyPDF2==3.0.0
python-docx==1.1.0
pandas==2.1.0

# For later phases (uncomment when needed)
# spacy==3.7.0
# sentence-transformers==2.2.0
# scikit-learn==1.3.0
# requests==2.31.0